hospital/requirements.txt
django==5.1
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
django-filter